Checkpoint: Move generate_solr_params_for_user since it is only used by user_files
