+++ refinery/user_files_manager/views.py
@@ -10,8 +10,9 @@ from django.template import RequestContext
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from data_set_manager.utils import (format_solr_response,
-                                    generate_solr_params_for_user, search_solr)
+from data_set_manager.utils import format_solr_response, search_solr
+
+from .utils import generate_solr_params_for_user
 
 logger = logging.getLogger(__name__)
